+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Initialize a tests suite
     - Use pwix:logger universal logger
     - Fix DataTables dt-container display style to be visible below dropdown menus
+    - Private Table._setCheckboxes() method is renamed _setTemplatesFromStrings() and is more extensible
 
 ### 1.6.0